Mega: fix build of new subsystems
Three issues from the previous commit:

1. Servo.h not found: PlatformIO's library finder didn't pull the
   Arduino-bundled Servo library by virtue of the include alone.
   Added 'Servo' to lib_deps so it gets resolved at build.

2. 'Wire' not declared in scope (hd44780_I2Cexp.h:378 and friends):
   the I2Cexp header uses Wire inline but does not include
   <Wire.h> itself. Added <Wire.h> before the I2Cexp include in
   i2c_lcd.h.

3. Sign-compare warning at i2c_lcd.cpp:77 (cols_ + 1 > sizeof(buf)):
   uint8_t plus int compared against size_t. Cast cols_ to size_t
   before the addition.

+++ firmware/io/src/subsystems/i2c_lcd.cpp
@@ -74,7 +74,7 @@ bool I2cLcd::handleSet(const proto::Command& cmd, Stream& out) {
     // Buffer caps at cols_+1; longer text gets truncated rather than
     // wrapping (we never want ambiguous overflow into the next row).
     char buf[33];   // covers up to 32 cols
-    if (cols_ + 1 > sizeof(buf)) {
+    if (static_cast<size_t>(cols_) + 1 > sizeof(buf)) {
       proto::writeError(out, "lcd.0", "geometry-too-wide");
       return true;
     }
